Address TUI Markdown review feedback
